third_party/WebKit/LayoutTests/fast/dom/SelectorAPI/duplicate-id-scope-expected.txt

Test that CSS selector queries on duplicate Ids respect scope.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.querySelectorAll('#duplicateId1').length is 6
PASS document.querySelectorAll('#duplicateId2').length is 2
PASS scopeDiv.querySelectorAll('#duplicateId1').length is 4
PASS scopeDiv.querySelector('#duplicateId2') is null
PASS scopeDiv.querySelectorAll('#duplicateId2').length is 0
PASS successfullyParsed is true

TEST COMPLETE
